server/requirements.txt
aniso8601==7.0.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-GraphQL==2.0.1
Flask-RESTful==0.3.10
Flask-Cors==5.0.1
Flask-Limiter==3.11.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
promise==2.3
pytz==2024.1
Rx==1.6.3
six==1.16.0
Werkzeug==3.0.3
zipp==3.19.2
pytest==8.3.5
pytest-cov==6.0.0
bcrypt==3.2.0
PyJWT==2.3.0
bleach==6.2.0
